Fixed bug, Post Category Template
Fixed a bug that was creating the "White Screen of Death" when attempting to access the single blog post template -> it was an unclosed if statement and while loop.  Also added a category template to list all the posts of a particular category however this is still a work in progress.